refactor(exceptions): Centralize orm/app errors and imports update
