+++ app/core/auth_utils.py
@@ -2,7 +2,7 @@ from werkzeug.security import generate_password_hash, check_password_hash
 from datetime import datetime, UTC, timedelta
 from app.model.models import User
 import jwt
-from app.core.errors_handlers import UnauthorizedError, BadRequestError
+from app.core.exceptions.app_errors import UnauthorizedError, BadRequestError
 from sqlalchemy.orm import Session
 from typing import Tuple
 from flask import current_app
