+++ app/services/product_utils.py
@@ -1,5 +1,5 @@
 from app.model.models import Product
-from app.core.errors_handlers import NotFoundError
+from app.core.exceptions.app_errors import NotFoundError
 from sqlalchemy.orm import Session
 from typing import List, Optional
